# apiserver/api_customization.go
/*
 * ABB Free@Home App API
 *
 * API to access and configure the ABB Free@Home App
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package apiserver

import (
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// CustomizationApiController binds http requests to an api service and writes the service results to the http response
type CustomizationApiController struct {
	service      CustomizationApiServicer
	errorHandler ErrorHandler
}

// CustomizationApiOption for how the controller is set up.
type CustomizationApiOption func(*CustomizationApiController)

// WithCustomizationApiErrorHandler inject ErrorHandler into controller
func WithCustomizationApiErrorHandler(h ErrorHandler) CustomizationApiOption {
	return func(c *CustomizationApiController) {
		c.errorHandler = h
	}
}

// NewCustomizationApiController creates a default api controller
func NewCustomizationApiController(s CustomizationApiServicer, opts ...CustomizationApiOption) Router {
	controller := &CustomizationApiController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the CustomizationApiController
func (c *CustomizationApiController) Routes() Routes {
	return Routes{
		{
			"GetDashboardTemplateByName",
			strings.ToUpper("Get"),
			"/v1/dashboard-templates/{dashboard-template-name}",
			c.GetDashboardTemplateByName,
		},
	}
}

// GetDashboardTemplateByName - Get a full dashboard template
func (c *CustomizationApiController) GetDashboardTemplateByName(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	query := r.URL.Query()
	dashboardTemplateNameParam := params["dashboard-template-name"]

	projectIdParam := query.Get("projectId")
	result, err := c.service.GetDashboardTemplateByName(r.Context(), dashboardTemplateNameParam, projectIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}